+++ packages/frontend/src/components/Explorer/ExploreTree/TableTree/Virtualization/flattenTree.ts
@@ -565,7 +565,11 @@ export function getNodeMapsForVirtualization(
         if (Object.keys(dimensionsMap).length > 0) {
             maps.set(
                 `${tableName}-dimensions`,
-                getNodeMapFromItemsMap(dimensionsMap, table.groupDetails),
+                getNodeMapFromItemsMap(
+                    dimensionsMap,
+                    table.groupDetails,
+                    table.orderFieldsBy,
+                ),
             );
         }
 
@@ -576,7 +580,11 @@ export function getNodeMapsForVirtualization(
         if (Object.keys(metricsMap).length > 0) {
             maps.set(
                 `${tableName}-metrics`,
-                getNodeMapFromItemsMap(metricsMap, table.groupDetails),
+                getNodeMapFromItemsMap(
+                    metricsMap,
+                    table.groupDetails,
+                    table.orderFieldsBy,
+                ),
             );
         }
 
@@ -590,7 +598,11 @@ export function getNodeMapsForVirtualization(
             );
             maps.set(
                 `${tableName}-custom-metrics`,
-                getNodeMapFromItemsMap(customMetricsMap, table.groupDetails),
+                getNodeMapFromItemsMap(
+                    customMetricsMap,
+                    table.groupDetails,
+                    table.orderFieldsBy,
+                ),
             );
         }
 
@@ -603,7 +615,11 @@ export function getNodeMapsForVirtualization(
             );
             maps.set(
                 `${tableName}-custom-dimensions`,
-                getNodeMapFromItemsMap(customDimensionsMap, table.groupDetails),
+                getNodeMapFromItemsMap(
+                    customDimensionsMap,
+                    table.groupDetails,
+                    table.orderFieldsBy,
+                ),
             );
         }
     });