+++ packages/frontend/src/components/common/ResourceView/index.tsx
@@ -140,7 +140,7 @@ const ResourceView: FC<ResourceViewProps> = ({
                                                                 icon={
                                                                     IconInfoCircle
                                                                 }
-                                                                color="ldGray.6"
+                                                                color="ldGray.9"
                                                             />
                                                         </Tooltip>
                                                     ) : null
@@ -148,7 +148,7 @@ const ResourceView: FC<ResourceViewProps> = ({
                                             >
                                                 {tab.name ? (
                                                     <Text
-                                                        color="ldGray.7"
+                                                        color="ldGray.9"
                                                         fz={15}
                                                         fw={500}
                                                     >